requirements.txt
bleach==3.1.4
bootstrap-admin==0.4.3
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.20
django-allauth==0.40.0
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-haystack==2.8.1
django-imagekit==4.0.2
django-redis==4.10.0
djangorestframework==3.10.3
django-user-agents==0.4.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jieba==0.39
Markdown==3.1.1
mysqlclient==1.4.4
oauthlib==3.1.0
olefile==0.46
pilkit==2.0
Pillow==6.2.0
Pygments==2.4.2
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
user-agent==0.1.9
webencodings==0.5.1
Whoosh==2.7.4
django-tctip>=1.4.0

python-dotenv~=0.19.2
geoip2~=4.5.0